Moving JDFTXInfile and JDFTXOutfile to inputs and outputs module, respectively
