+++ tests/io/jdftx/test_jdftxinfile.py
@@ -8,7 +8,7 @@ import numpy as np
 import pytest
 
 from pymatgen.core.structure import Structure
-from pymatgen.io.jdftx.jdftxinfile import JDFTXInfile, JDFTXStructure
+from pymatgen.io.jdftx.inputs import JDFTXInfile, JDFTXStructure
 from pymatgen.io.jdftx.jdftxinfile_master_format import get_tag_object
 
 from .conftest import assert_same_value, dump_files_dir, ex_files_dir
